fix: refactor Topnav
