+++ src/components/index.tsx
@@ -1,0 +1,2 @@
+export {default as HeaderComponent} from './HeaderComponent';
+export {default as NavItem} from './NavItem';
